fixed int/float conversion

+++ sw/matrix_video_player.py
@@ -19,7 +19,7 @@ color_scale = 8   # color divider
 num_rows = 64
 num_cols = num_rows
 output_size = num_rows
-bin_size = input_size / output_size
+bin_size = int(input_size / output_size)
 fbuf = np.zeros((num_rows), dtype='i4')
 
 s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
@@ -34,7 +34,7 @@ while(1):
     for y in range(num_rows):
         for x in range(num_cols):
             addr = ((y & 0x3F) << 6) | (x & 0x3F);
-            
+
             r = cast[y][x][2].item()
             g = cast[y][x][0].item()
             b = cast[y][x][1].item()
